Move the modal for adding MUCs to its own file
